pyproject.toml
[tool.poetry]
name = "onlinecinemaapi"
version = "0.1.0"
description = "Online Cinema API"
authors = ["Team"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.115.7,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
sqlalchemy = ">=2.0.37,<3.0.0"
alembic = ">=1.14.1,<2.0.0"
psycopg = {extras = ["binary"], version = ">=3.2.4,<4.0.0"}
pydantic = ">=2.10.6,<3.0.0"
pydantic-settings = ">=2.7.1,<3.0.0"
passlib = "^1.7.4"
email-validator = "^2.2.0"
uuid = "^1.30"
pycryptodome = "^3.21.0"
bcrypt = "4.0.1"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
stripe = "^11.5.0"
pillow = "^11.1.0"
boto3 = "^1.36.11"
botocore = "^1.36.11"
python-multipart = "^0.0.20"
requests = "^2.32.3"
jinja2 = "^3.1.5"
slowapi = "^0.1.9"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
celery = {extras = ["beat"], version = "^5.4.0"}
redis = "^5.2.1"
pytest = "^8.3.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
isort = "^5.13.2"
coverage = "^7.6.10"

[tool.ruff]
line-length = 120
exclude = [
    "migrations",
    "node_modules",
    "static",
    "templates",
    "venv",
    "env",
    ".git",
    ".idea",
    ".venv",
    ".dockerignore",
    ".env",
    ".gitignore",
    ".classpath",
    ".DS_Store",
    "*.dbmdl",
    "*.user",
    "*.dockerignore",
    "*.env",
    "*.git",
    "*.gitignore",
    "*.project",
    "*.settings",
    "*.toolstarget",
    "*.classpath",
    "*.DS_Store",
]

[tool.ruff.lint]
select = [
    "E",  # PEP8 errors
    "W",  # PEP8 warnings
    "F",  # PyFlakes
    "I",  # isort
    "B",  # bugbear
    "C4",  # mypy
    "N",  # flake8-bugbear
    "Q",  # flake8-quotes
]
ignore = [
    "C901",  # function too complex
    "W191",  # indentation contains tabs
    "D401",  # first line should be in imperative mood
    "E712",
    "I001",   # isort
    "F401",
    "F821",
    "B008",
    "B904",
    "N802",
    "Q000"
]

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
lines_after_imports = 2